This branch goes over some of the Read(GET) and DELETE components of CRUD
In authorData.js(under the api folder), I made two functions that have promises in them, one function to GET all authors and one function to DELETE a singular author. In bookData.js(under the api folder), I made three functions that have promises in them. One function is to GET all books, one function to DELETE a singular book, and the last function is to filter(GET) books by favorite status. In domEvents.js(under the events folder), I made two if statements underneath the querySelector Event Listener, which is called through the domEvents function. One if statement pertained to deleting books, and the other statement was for deleting an author. In navigationEvents.js (under the events folder), I added three callback functions attached to .then() methods, which can all be found underneath their own associated querySelector Event Listener, which can be called through the navigationEvents function. One callback function focuses on displaying(navigating to) all of the books that are on sale, the second callback function focuses on displaying(navigating to) all of the books, and the third callback function focuses on displaying(navigating to) all of the authors. Lastly, in startApp.js(underneath the utils folder), I added the function to have all of the books on the DOM. This process first calls the function getBooks(), which will eventually produce a promise. When getBooks produces that promise, then it will run this code with the result of books. When those two parts of the process happen, lastly, showBooks will will show all of the current books from the firebase data.